Add files via upload
The Titanic dataset is a well-known and widely used dataset in the field of data science and machine learning. It is one of the most popular datasets for beginners due to its relative simplicity and historical significance. The dataset contains information about the passengers aboard the RMS Titanic, which famously sank on its maiden voyage in April 1912 after colliding with an iceberg. The dataset provides valuable insights into the passengers' demographics and survival status.

The features (columns) in the Titanic dataset typically include:

PassengerId: A unique identifier for each passenger.
Survived: This is the target variable (dependent variable) that indicates whether the passenger survived (1) or not (0).
Pclass (Passenger Class): The ticket class of the passenger, representing their socio-economic status. It can take values 1, 2, or 3 (1st, 2nd, or 3rd class).
Name: The name of the passenger.
Sex: The gender of the passenger (male or female).
Age: The age of the passenger in years. It can contain missing values.
SibSp: The number of siblings/spouses aboard the Titanic.
Parch: The number of parents/children aboard the Titanic.
Ticket: The ticket number.
Fare: The fare paid for the ticket.
Cabin: The cabin number of the passenger. This feature may have many missing values.
Embarked: The port of embarkation (C = Cherbourg, Q = Queenstown, S = Southampton).
The main objective with this dataset is to predict whether a passenger survived the disaster based on the given features. It allows data analysts and machine learning practitioners to explore various data preprocessing techniques, feature engineering methods, and predictive models to better understand the factors that might have influenced a passenger's chances of survival.

Researchers and data enthusiasts have used this dataset to develop and compare different machine learning algorithms, classification models, and survival prediction techniques. It serves as an accessible and practical resource for individuals learning about data analysis, data visualization, and predictive modeling in the context of a historical event with significant human interest.